requirements.txt
alembic==1.0.10
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.6.0
dominate==2.3.5
elasticsearch==7.0.2
entrypoints==0.3
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.8.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Mako==1.0.12
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
pipp==0.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pyrsistent==0.15.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.5
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
WTForms==2.2.1
